[Refactor] useCase.sectionList 구독 공유로 createTimeTableCellList 중복 호출 제거
