+++ Projects/Features/TimeTableFeature/Sources/View/BottomSheet/Search/SearchModuleView.swift
@@ -64,13 +64,13 @@ public struct SearchModuleView: View {
                 
             } else {
                 ScrollView {
-                    ForEach(viewModel.lectureList, id: \.self) { lecture in
+                    ForEach(viewModel.lectureList) { lecture in
                         ClassSearchListCellView(
                             isSelected: viewModel.state.selectedLecture == lecture,
                             section: lecture,
                             cellDidTap: { viewModel.send(.lectureCellDidTap(lecture)) },
                             addLectureDidTap: { viewModel.send(.addLectureButtonDidTap(lecture))}
-                        )
+                            )
                         .padding(.bottom, 3)
                     }
                 }
