+++ Projects/Features/TimeTableFeature/Sources/View/Main/TimeTableExsitedView.swift
@@ -1,0 +1,155 @@
+import SwiftUI
+import Domain
+import DSKit
+
+public struct TimeTableExsitedView: View {
+    @ObservedObject var viewModel: TimeTableViewModel
+    @Binding var displayType: DisplayTypeInfo
+    @Binding var viewType: TimeTableViewType
+    var cellWidth: CGFloat
+    
+    init(
+        viewModel: TimeTableViewModel,
+        displayType: Binding<DisplayTypeInfo>,
+        viewType: Binding<TimeTableViewType>,
+        cellWidth: CGFloat
+    ) {
+        self.viewModel = viewModel
+        self._displayType = displayType
+        self._viewType = viewType
+        self.cellWidth = cellWidth
+    }
+    
+    public var body: some View {
+        // 📌 수업 버튼 배치
+        ForEach($viewModel.timeTable, id: \.self) { $cell in
+            if let dayIndex = viewModel.weekList.firstIndex(of: cell.schedule.day) {
+                let rect: (centerX: CGFloat, centerY: CGFloat, height: CGFloat) = configButtonLayout(
+                    viewModel.hourList[0],
+                    for: cell,
+                    at: dayIndex,
+                    cellWidth: cellWidth,
+                    cellHeight: 52
+                )
+                
+                let backgroundColor: Color = viewModel.selectedThemeColor.isEmpty
+                ? cell.backgroundColor
+                : Color.init(hex: viewModel.selectedThemeColor.randomElement()!)
+                
+                ZStack {
+                    createClassButton(
+                        for: cell,
+                        centerX: rect.centerX,
+                        centerY: rect.centerY,
+                        cellWidth: cellWidth,
+                        cellHeight: rect.height,
+                        backgroundColor: backgroundColor
+                    )
+                    
+                    let textColor: Color = viewModel.selectedThemeColor.isEmpty
+                    ? cell.textColor
+                    : Color.init(hex: viewModel.selectedThemeColor[0])
+                    
+                    createClassInfoText(
+                        for: cell,
+                        textColor: textColor,
+                        centerX: rect.centerX,
+                        centerY: rect.centerY,
+                        cellWidth: cellWidth,
+                        cellHeight: rect.height
+                    )
+                }
+                .onTapGesture {
+                    viewModel.send(.tableCellDidTap(cell.id))
+                }
+                .disabled(viewType != .main)
+            }
+        }
+    }
+}
+extension TimeTableExsitedView {
+    private func createClassButton(
+        for cell: TimeTableCellInfo,
+        centerX: CGFloat,
+        centerY: CGFloat,
+        cellWidth: CGFloat,
+        cellHeight: CGFloat,
+        backgroundColor: Color
+    ) -> some View {
+        return Button {
+            withAnimation {
+                viewModel.send(.tableCellDidTap(cell.id))
+            }
+        } label: {
+            Rectangle()
+                .fill(backgroundColor)
+                .clipShape(RoundedRectangle(cornerRadius: 2))
+                .frame(width: cellWidth-1, height: cellHeight-1)
+                .position(x: centerX, y: centerY)
+        }
+        .buttonStyle(PlainButtonStyle())
+    }
+    
+    
+    private func createClassInfoText(
+        for cell: TimeTableCellInfo,
+        textColor: Color,
+        centerX: CGFloat,
+        centerY: CGFloat,
+        cellWidth: CGFloat,
+        cellHeight: CGFloat
+    ) -> some View {
+        return VStack(alignment: .leading, spacing: 0) {
+            Text(cell.code)
+                .font(.medium_12)
+                .foregroundColor(textColor)
+                .multilineTextAlignment(.leading)
+                .lineLimit(1)
+            
+            if displayType.classRoomIsVisible {
+                Text(cell.schedule.location)
+                    .font(.regular_10)
+                    .foregroundColor(textColor)
+                    .lineLimit(1)
+            }
+            
+            if displayType.creditIsVisible, let unit = cell.unit {
+                Text("unit: \(unit)")
+                    .font(.regular_10)
+                    .foregroundColor(textColor)
+                    .lineLimit(1)
+            }
+        }
+        .frame(width: 56, height: cellHeight ,alignment: .topLeading)
+        .position(x: centerX-4, y: centerY)
+    }
+}
+
+extension TimeTableExsitedView {
+    private func configButtonLayout(
+        _ firstTime: Int,
+        for cell: TimeTableCellInfo,
+        at dayIndex: Int,
+        cellWidth: CGFloat,
+        cellHeight: CGFloat
+    ) -> (centerX: CGFloat, centerY: CGFloat, height: CGFloat) {
+        let startHour = cell.schedule.startHour
+        let startMinute = cell.schedule.startMinute
+        let endHour = cell.schedule.endHour
+        let endMinute = cell.schedule.endMinute
+        
+        
+        let x = CGFloat(dayIndex) * cellWidth
+        // 시작 시간과 분을 기준으로 시작 위치 계산
+        let y = CGFloat(startHour - firstTime) * cellHeight + CGFloat(startMinute) / 60 * cellHeight
+        
+        // 종료 시간과 분을 기준으로 높이 계산
+        let height = CGFloat(endHour - startHour) * cellHeight +
+        CGFloat(endMinute - startMinute) / 60 * cellHeight
+        
+        let centerX = x + cellWidth / 2
+        let centerY =  y + height / 2 + cellHeight / 2
+        return (centerX, centerY, height)
+    }
+}
+
